feat: Update lastModified and createdAt timestamps in global memory bank tags JSON
