+++ docs/global-memory-bank/tags/index.json
@@ -9,8 +9,8 @@
       "index",
       "meta"
     ],
-    "lastModified": "2025-03-30T10:11:21.674Z",
-    "createdAt": "2025-03-30T10:11:21.674Z",
+    "lastModified": "2025-03-30T10:19:56.697Z",
+    "createdAt": "2025-03-30T10:19:56.697Z",
     "version": 1
   },
   "content": {
